requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3
Django==1.10
django-celery==3.2.2
django-cors-headers==2.4.0
django-extensions==2.1.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.8.2
enum34==1.1.6
idna==2.7
ipaddress==1.0.22
itypes==1.1.0
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
mysql-connector-python==8.0.12
MySQL-python==1.2.5
openapi-codec==1.3.2
Pillow==5.2.0
protobuf==3.6.0
pycparser==2.18
pycrypto==2.6.1
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.19.1
simple-crypt==4.1.7
simple-salesforce==0.74.2
simplejson==3.16.0
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
